move auto scaling test to tidbcluster
